+++ src/main/java/com/WorkoutHub/workout_hub/dto/response/ExerciseHistoryDto.java
@@ -3,6 +3,7 @@ package com.WorkoutHub.workout_hub.dto.response;
 
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -17,7 +18,7 @@ import java.util.List;
 @AllArgsConstructor
 public class ExerciseHistoryDto {
     @JsonProperty("exercise_info_id")
-    private Long exerciseInfoId;
+    private Integer exerciseInfoId;
 
     @JsonProperty("exercise_name")
     private String exerciseName;
